requirements.txt
pypdf
langchain
langchain_community
langchain-chroma
chromadb # Vector storage
pytest
boto3
langchain-aws
pdf2image
Pillow